Update grid.sect.R
working directly on git-hub today. I tried to run this function to better understand the values and I receive an error on the while(h-l>.0000001 line. It states that we are missing a True/False statement. I am not sure what it is asking for, can you please review it Tyler? I also removed @ return and added tyler as author.